refactor launchers
