+++ scripts/iterm.sh
@@ -1,23 +1,10 @@
-CMD=""
-WD="$PWD"
-ARGS="$@"
-
-if [ -d "$1" ]; then
-    WD="$1"
-    ARGS="${@:2}"
-fi
-
-if [ -n "$ARGS" ]; then
-    CMD="$ARGS"
-fi
-
-    osascript <<EOF
+osascript <<EOF
 tell application "iTerm2"
   tell current session of current window
-    set new_session to split $SPLIT_DIRECTION with default profile
+    set new_session to split $__SPLIT_DIRECTION__ with same profile
       tell new_session
         delay 1
-		    write text "cd $WD;$CMD"
+		    write text "cd $__CD__;$__CMD__"
       end tell
   end tell
 end tell
