+++ scripts/tmux.sh
@@ -1,18 +1,5 @@
-CMD=""
-WD="$PWD"
-ARGS="$@"
-
-if [ -d "$1" ]; then
-    WD="$1"
-    ARGS="${@:2}"
-fi
-
-if [ -n "$ARGS" ]; then
-    CMD="$ARGS"
-fi
-
 currentPane=$(tmux display-message -p '#P')
 newPane=$(($currentPane + 1))
-tmux split-window -$SPLIT_DIRECTION
+tmux split-window -$__SPLIT_DIRECTION__
 tmux select-pane -t "$currentPane"
-tmux send-keys -t "$newPane" C-z "cd $WD;$CMD" Enter
+tmux send-keys -t "$newPane" C-z "cd $__CD__;$__CMD__" Enter
